Fade out old slide, then fade in new slide
Previously, would fade out old slide and fade in new slide at exact
same time. Now old slide is faded out and when that animation completes
the new slide will begin fading in.

Also added some CSS to make the slide viewer UL's display property
block before animating. This allows us to set the  property for it to
none on load, effectively making it so that the user doesn't see all of
the slides spilling out of the slide viewer before the JS loads to hide
them.